+++ .dev/dev-constants.js
@@ -1,7 +1,7 @@
 /*!
 * Constants
 * Description: Constants for the development.
-* Version: 2.2.0
+* Version: 2.3.0
 * Author: Furkan ('https://github.com/furcan')
 * Copyright 2022 Constants, MIT Licence ('https://opensource.org/licenses/MIT')
 */